Rename BrandingInfo to BrandingHeader
